+++ src/atomsWithTanstackQuery.ts
@@ -3,6 +3,7 @@ import type {
   QueryKey,
   QueryObserverOptions,
   QueryObserverResult,
+  ResetOptions,
 } from '@tanstack/query-core'
 import { atom } from 'jotai'
 import type { Getter } from 'jotai'
@@ -11,6 +12,8 @@ import { queryClientAtom } from './queryClientAtom'
 
 type Action = {
   type: 'refetch'
+  force?: boolean
+  options?: ResetOptions
 }
 
 export function atomsWithTanstackQuery<
@@ -23,16 +26,32 @@ export function atomsWithTanstackQuery<
   getOptions: (
     get: Getter
   ) => QueryObserverOptions<TQueryFnData, TError, TData, TQueryData, TQueryKey>,
-  getQueryClient?: (get: Getter) => QueryClient
+  getQueryClient: (get: Getter) => QueryClient = (get) => get(queryClientAtom)
 ) {
   type Result = QueryObserverResult<TData, TError>
 
+  const observerCacheAtom = atom(
+    () =>
+      new WeakMap<
+        QueryClient,
+        QueryObserver<TQueryFnData, TError, TData, TQueryData, TQueryKey>
+      >()
+  )
+
+  const refreshAtom = atom(0)
+
   const observerAtom = atom((get) => {
-    const queryClient = getQueryClient
-      ? getQueryClient(get)
-      : get(queryClientAtom)
+    get(refreshAtom)
+    const queryClient = getQueryClient(get)
     const options = getOptions(get)
-    const observer = new QueryObserver(queryClient, options)
+    const observerCache = get(observerCacheAtom)
+    let observer = observerCache.get(queryClient)
+    if (observer) {
+      observer.setOptions(options)
+    } else {
+      observer = new QueryObserver(queryClient, options)
+      observerCache.set(queryClient, observer)
+    }
     return observer
   })
 
@@ -61,10 +80,18 @@ export function atomsWithTanstackQuery<
 
   const statusAtom = atom(
     (get) => get(baseStatusAtom),
-    (get, _set, action: Action) => {
+    (get, set, action: Action) => {
       if (action.type === 'refetch') {
         const observer = get(observerAtom)
-        return observer.refetch({ cancelRefetch: true }).then(() => {})
+        if (action.force) {
+          observer.remove()
+          const queryClient = getQueryClient(get)
+          const observerCache = get(observerCacheAtom)
+          observerCache.delete(queryClient)
+          set(refreshAtom, (c) => c + 1)
+          return
+        }
+        return observer.refetch(action.options).then(() => {})
       }
     }
   )
@@ -96,12 +123,7 @@ export function atomsWithTanstackQuery<
       }
       return baseData.data as TData
     },
-    (get, _set, action: Action) => {
-      if (action.type === 'refetch') {
-        const observer = get(observerAtom)
-        return observer.refetch({ cancelRefetch: true }).then(() => {})
-      }
-    }
+    (_get, set, action: Action) => set(statusAtom, action)
   )
 
   return [dataAtom, statusAtom] as const