+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - implement refetch #1
+- feat: observer cache #2
 
 ## [0.0.1] - 2022-09-20
 ### Added